Pieces for agents/entity mostly there. Almost all uarmtd_agent functionality is theoretically transferred over. Next is glue

+++ simulation/agent/EntityInfo.m
@@ -1,7 +1,7 @@
 classdef EntityInfo < handle
     % EntityInfo This might go later
     properties (Abstract)
-        dimension uint8 {mustBeMember(dimension,[2, 3])}
+        dimension double {mustBeMember(dimension,[2, 3])}
     end
     methods (Abstract)
         reset(self)
